refactor: separate oauth, task store and response helpers
