+++ pkg/api/sync_handler.go
@@ -1,7 +1,6 @@
 package api
 
 import (
-	"database/sql"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -11,7 +10,6 @@ import (
 
 	"github.com/go-chi/chi/v5"
 	"github.com/retail-ai-inc/sync/internal/db/mongodb"
-	"github.com/retail-ai-inc/sync/pkg/db"
 	"github.com/sirupsen/logrus"
 )
 
@@ -541,99 +539,4 @@ func SyncTablesHandler(w http.ResponseWriter, r *http.Request) {
 			"tables":     tableStats,
 		},
 	})
-}
-
-// -------------------------
-// Helpers & Common Utilities
-// -------------------------
-
-func openLocalDB() (*sql.DB, error) {
-	return db.OpenSQLiteDB()
-}
-
-func updateTaskStatus(id string, toStart bool) error {
-	db, err := openLocalDB()
-	if err != nil {
-		return err
-	}
-	defer db.Close()
-
-	var oldCfgJSON string
-	err = db.QueryRow(`SELECT config_json FROM sync_tasks WHERE id=?`, id).Scan(&oldCfgJSON)
-	if err != nil {
-		return err
-	}
-	var statusStr string
-	var newEnable int
-	if toStart {
-		statusStr = "Running"
-		newEnable = 1
-	} else {
-		statusStr = "Stopped"
-		newEnable = 0
-	}
-
-	var data map[string]interface{}
-	if err2 := json.Unmarshal([]byte(oldCfgJSON), &data); err2 != nil {
-		data = make(map[string]interface{})
-	}
-	data["status"] = statusStr
-	newBytes, _ := json.Marshal(data)
-
-	nowStr := timeNowStr()
-	_, err = db.Exec(`
-UPDATE sync_tasks
-SET enable=?,
-    last_update_time=?,
-    config_json=?
-WHERE id=?
-`, newEnable, nowStr, string(newBytes), id)
-	return err
-}
-
-func errorJSON(w http.ResponseWriter, msg string, err error) {
-	logrus.Errorf("%s => %v", msg, err)
-	resp := map[string]interface{}{
-		"success": false,
-		"error":   msg,
-		"detail":  err.Error(),
-	}
-	writeJSON(w, resp)
-}
-
-func writeJSON(w http.ResponseWriter, data interface{}) {
-	w.Header().Set("Content-Type", "application/json")
-	_ = json.NewEncoder(w).Encode(data)
-}
-
-// timeNowStr returns the current time formatted as a string in UTC timezone
-// for database storage purposes
-func timeNowStr() string {
-	return time.Now().UTC().Format("2006-01-02 15:04:05")
-}
-
-// convertTimeToJST converts a time string from UTC to JST timezone for SQL time format
-// This handles the specific format used in the database "2006-01-02 15:04:05"
-func convertTimeToJST(input string) string {
-	if input == "" {
-		return ""
-	}
-
-	// First try parsing with standard SQL format
-	layout := "2006-01-02 15:04:05"
-	t, err := time.Parse(layout, input)
-	if err == nil {
-		jst := time.FixedZone("JST", 9*60*60)
-		return t.In(jst).Format(layout)
-	}
-
-	// If that fails, try RFC3339 format
-	t, err = time.Parse(time.RFC3339, input)
-	if err == nil {
-		jst := time.FixedZone("JST", 9*60*60)
-		return t.In(jst).Format(layout)
-	}
-
-	// Return original if we can't parse
-	return input
 }
